Создал абстрактный класс BaseFragment.kt с настройкой binding для наследуемых классов-фрагментов.
