+++ app/src/main/java/com/example/nasaapp/view/ChoosingTheDayFragment.kt
@@ -6,10 +6,7 @@ import android.net.Uri
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
 import com.example.nasaapp.R
 import com.example.nasaapp.databinding.ChoosingTheDayLayoutBinding
@@ -17,13 +14,11 @@ import com.example.nasaapp.model.dto.AstronomyPictureOfTheDay
 import com.example.nasaapp.utils.*
 import com.google.android.material.tabs.TabLayoutMediator
 
-class ChoosingTheDayFragment:Fragment() {
+class ChoosingTheDayFragment:BaseFragment<ChoosingTheDayLayoutBinding>(ChoosingTheDayLayoutBinding::inflate) {
+
     companion object {
         fun newInstance() = ChoosingTheDayFragment()
     }
-
-    private var _binding: ChoosingTheDayLayoutBinding? = null
-    private val binding get() = _binding!!
 
     private val listAstronomyPictureOfTheDay: MutableList<AstronomyPictureOfTheDay?> = mutableListOf<AstronomyPictureOfTheDay?>().apply {
         Day.values().forEach { _ ->
@@ -33,15 +28,6 @@ class ChoosingTheDayFragment:Fragment() {
 
     private lateinit var viewPager: ViewPager2
     private lateinit var adapterForAstronomyPicturesOfTheDay: ViewPagerAdapterForAstronomyPicturesOfTheDay
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = ChoosingTheDayLayoutBinding.inflate(inflater)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -178,11 +164,6 @@ class ChoosingTheDayFragment:Fragment() {
     override fun onPause() {
         saveChoosingDay(Day.values()[viewPager.currentItem])
         super.onPause()
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
 }
