+++ app/src/main/java/com/example/nasaapp/view/MarsRoverPhotosRecyclerViewFragment.kt
@@ -1,10 +1,7 @@
 package com.example.nasaapp.view
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
@@ -12,15 +9,12 @@ import com.example.nasaapp.databinding.MarsPhotosRecyclerViewFragmentBinding
 import com.example.nasaapp.view_model.MarsRoverPhotosViewModel
 
 
-class MarsRoverPhotosRecyclerViewFragment : Fragment() {
+class MarsRoverPhotosRecyclerViewFragment : BaseFragment<MarsPhotosRecyclerViewFragmentBinding>(MarsPhotosRecyclerViewFragmentBinding::inflate) {
 
     companion object {
         fun newInstance(): MarsRoverPhotosRecyclerViewFragment =
             MarsRoverPhotosRecyclerViewFragment()
     }
-
-    private var _binding: MarsPhotosRecyclerViewFragmentBinding? = null
-    private val binding get() = _binding!!
 
    private lateinit var adapter: AdapterForMarsRoverPhotosRecyclerViewFragment
 
@@ -93,14 +87,6 @@ class MarsRoverPhotosRecyclerViewFragment : Fragment() {
         ViewModelProvider(this)[MarsRoverPhotosViewModel::class.java]
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-       _binding = MarsPhotosRecyclerViewFragmentBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         adapter = AdapterForMarsRoverPhotosRecyclerViewFragment(viewModel.getListPhoto(),callback)
@@ -108,8 +94,4 @@ class MarsRoverPhotosRecyclerViewFragment : Fragment() {
         ItemTouchHelper(ItemTouchHelperCallback(adapter)).attachToRecyclerView(binding.marsPhotosRecyclerView)
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
