+++ app/src/main/java/com/example/nasaapp/view/ViewPagerForMarsRoverPhotosFragment.kt
@@ -2,10 +2,7 @@ package com.example.nasaapp.view
 
 import android.os.Bundle
 import android.util.Log
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.viewpager2.widget.ViewPager2
 import com.example.nasaapp.databinding.ViewPagerForMarsRoverPhotosFragmentBinding
@@ -15,16 +12,13 @@ import com.example.nasaapp.view_model.AppStateMarsRoverPhotos
 import com.example.nasaapp.view_model.MarsRoverPhotosViewModel
 
 
-class ViewPagerForMarsRoverPhotosFragment : Fragment() {
+class ViewPagerForMarsRoverPhotosFragment : BaseFragment<ViewPagerForMarsRoverPhotosFragmentBinding> (ViewPagerForMarsRoverPhotosFragmentBinding::inflate) {
 
     companion object {
         fun newInstance(): ViewPagerForMarsRoverPhotosFragment =
             ViewPagerForMarsRoverPhotosFragment()
         val day = Day.DAY_BEFORE_YESTERDAY
     }
-
-    private var _binding: ViewPagerForMarsRoverPhotosFragmentBinding? = null
-    private val binding get() = _binding!!
 
     private val viewModel: MarsRoverPhotosViewModel by lazy {
         ViewModelProvider(this)[MarsRoverPhotosViewModel::class.java]
@@ -34,14 +28,6 @@ class ViewPagerForMarsRoverPhotosFragment : Fragment() {
     private lateinit var adapterForMarsRoverPhotos:ViewPagerAdapterForMarsRoverPhotos
 
      private var marsRoverPhotos: MarsRoverPhotos?=null
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = ViewPagerForMarsRoverPhotosFragmentBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -116,8 +102,4 @@ class ViewPagerForMarsRoverPhotosFragment : Fragment() {
         adapterForMarsRoverPhotos.notifyItemChanged(viewPager.currentItem)
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
